requirements.txt
absl-py==2.1.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
gunicorn==21.2.0
idna==3.10
immutabledict==4.2.1
iniconfig==2.0.0
mailjet-rest==1.3.4
numpy==2.2.3
openpyxl==3.1.5
ortools==9.12.4544
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
protobuf==5.29.3
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyMuPDF==1.25.4
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
